+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.1.0] - 2024-04-26
+### Added
+- Added additional data fields
+
+### Fixed
+- Fixed extraction of `Total Reviews` and `Website` data fields
+
+### Changed
+- Updated logic of parsing data fields
+- Changed output data file name
+
 ## [3.0.0] - 2024-04-26
 
 ### Fixed